move deleteFeedback and addFeedback functions to Context Provider to simplify component instances. Global state, and state mutator funtions are now combined in FeedbackContext.js
