+++ src/App.js
@@ -1,9 +1,6 @@
-import { useState } from 'react';
-import { v4 as uuidv4 } from 'uuid';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Header from './components/Header';
 import FeedbackList from './components/FeedbackList';
-import FeedbackData from './data/FeedbackData';
 import FeedbackStats from './components/FeedbackStats';
 import FeedbackForm from './components/FeedbackForm';
 import AboutPage from './pages/AboutPage';
@@ -12,18 +9,6 @@ import { FeedbackProvider } from './context/FeedbackContext';
 
 const App = () => {
     // Create state
-    const [feedback, setFeedback] = useState(FeedbackData);
-
-    const deleteFeedback = (id) => {
-        if (window.confirm('Are you sure you want to delete this feedback?')) {
-            setFeedback(feedback.filter((item) => item.id !== id));
-        }
-    };
-
-    const addFeedback = (newFeedback) => {
-        newFeedback.id = uuidv4();
-        setFeedback([newFeedback, ...feedback]);
-    };
 
     return (
         <FeedbackProvider>
@@ -36,11 +21,9 @@ const App = () => {
                             path="/"
                             element={
                                 <>
-                                    <FeedbackForm handleAdd={addFeedback} />
+                                    <FeedbackForm />
                                     <FeedbackStats />
-                                    <FeedbackList
-                                        handleDelete={deleteFeedback}
-                                    />
+                                    <FeedbackList />
                                     <AboutBtnLink />
                                 </>
                             }
