+++ src/components/FeedbackItem.jsx
@@ -1,13 +1,20 @@
 import { FaTimes } from 'react-icons/fa';
 import Card from './shared/Card';
 import PropTypes from 'prop-types';
+import { useContext } from 'react';
+import FeedbackContext from '../context/FeedbackContext';
 
-function FeedbackItem({ item, handleDelete }) {
+function FeedbackItem({ item }) {
+    const { deleteFeedback } = useContext(FeedbackContext);
+
     return (
         <Card>
             <div className="num-display">{item.rating}</div>
             <button className="close">
-                <FaTimes onClick={() => handleDelete(item.id)} color="purple" />
+                <FaTimes
+                    onClick={() => deleteFeedback(item.id)}
+                    color="purple"
+                />
             </button>
             <div className="text-display">{item.text}</div>
         </Card>
