+++ src/components/FeedbackList.jsx
@@ -4,7 +4,7 @@ import FeedbackItem from './FeedbackItem';
 import FeedbackContext from '../context/FeedbackContext';
 
 // Component for entire list of feedback cards
-const FeedbackList = ({ handleDelete }) => {
+const FeedbackList = () => {
     const { feedback } = useContext(FeedbackContext);
 
     if (!feedback || feedback.length === 0) {
@@ -23,11 +23,7 @@ const FeedbackList = ({ handleDelete }) => {
                         animate={{ opacity: 1 }}
                         exit={{ opacity: 0 }}
                     >
-                        <FeedbackItem
-                            key={item.id}
-                            item={item}
-                            handleDelete={handleDelete}
-                        />
+                        <FeedbackItem key={item.id} item={item} />
                     </motion.div>
                 ))}
             </AnimatePresence>
