+++ src/context/FeedbackContext.js
@@ -1,4 +1,5 @@
 import { createContext, useState } from 'react';
+import { v4 as uuidv4 } from 'uuid';
 
 const FeedbackContext = createContext();
 
@@ -6,13 +7,41 @@ export const FeedbackProvider = ({ children }) => {
     const [feedback, setFeedback] = useState([
         {
             id: 1,
-            text: 'This item is from context',
+            text: 'This is feedback item 1',
             rating: 10,
+        },
+        {
+            id: 2,
+            text: 'This is feedback item 2',
+            rating: 2,
+        },
+        {
+            id: 3,
+            text: 'This is feedback item 3',
+            rating: 9,
+        },
+        {
+            id: 4,
+            text: 'This is feedback item 4',
+            rating: 7,
         },
     ]);
 
+    const deleteFeedback = (id) => {
+        if (window.confirm('Are you sure you want to delete this feedback?')) {
+            setFeedback(feedback.filter((item) => item.id !== id));
+        }
+    };
+
+    const addFeedback = (newFeedback) => {
+        newFeedback.id = uuidv4();
+        setFeedback([newFeedback, ...feedback]);
+    };
+
     return (
-        <FeedbackContext.Provider value={{ feedback }}>
+        <FeedbackContext.Provider
+            value={{ feedback, deleteFeedback, addFeedback }}
+        >
             {children}
         </FeedbackContext.Provider>
     );
